requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
alembic==1.13.3
pandas==2.2.3
python-multipart==0.0.9
redis==5.0.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
